Refactor dump_ndjson() and load_ndjson() into id3c.json
For use in other parts of the codebase, which right now means our
id3c-customizations.
